+++ tests/test_graph.py
@@ -3,3 +3,4 @@ import libsignal
 def test_graph():
     graph = libsignal.AudioGraph()
     assert graph
+    del graph